refactored slack dialog to be in the helper slack module
